requirements.txt
Django==2.2.2
django-cors-headers==3.0.2
httpie==0.9.8
httplib2==0.9.2
numpy==1.16.2
oauth==1.0.1
pandas==0.24.2
djangorestframework==3.9.4
django-picklefield==2.0
